+++ gp2gp-translator/src/test/java/uk/nhs/adaptors/pss/translator/mapper/medication/MedicationRequestPlanMapperTest.java
@@ -2,15 +2,14 @@ package uk.nhs.adaptors.pss.translator.mapper.medication;
 
 import static org.hl7.fhir.dstu3.model.MedicationRequest.MedicationRequestStatus.COMPLETED;
 import static org.hl7.fhir.dstu3.model.MedicationRequest.MedicationRequestStatus.STOPPED;
+import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
-import static org.springframework.util.ResourceUtils.getFile;
 import static org.assertj.core.api.Assertions.assertThat;
-
-import static uk.nhs.adaptors.pss.translator.util.XmlUnmarshallUtil.unmarshallFile;
 
 import static org.hl7.fhir.dstu3.model.MedicationRequest.MedicationRequestStatus.ACTIVE;
 import static org.hl7.fhir.dstu3.model.MedicationRequest.MedicationRequestIntent.PLAN;
+import static uk.nhs.adaptors.pss.translator.util.XmlUnmarshallUtil.unmarshallString;
 
 import java.util.List;
 import java.util.Optional;
@@ -45,7 +44,22 @@ import uk.nhs.adaptors.pss.translator.util.DateFormatUtil;
 @ExtendWith(MockitoExtension.class)
 public class MedicationRequestPlanMapperTest {
 
-    private static final String XML_RESOURCES_BASE = "xml/MedicationStatement/";
+    private static final String EHR_EXTRACT_WRAPPER = """
+        <EhrExtract xmlns="urn:hl7-org:v3">
+            <component>
+                <ehrFolder>
+                    <component>
+                        <ehrComposition>
+                            <component>
+                                {{MedicationStatement}}
+                            </component>
+                        </ehrComposition>
+                    </component>
+                </ehrFolder>
+            </component>
+        </EhrExtract>
+        """;
+
     private static final String PRACTISE_CODE = "TESTPRACTISECODE";
     private static final String MEDICATION_ID = "MEDICATION_ID";
     private static final String TEST_ID = "TEST_ID";
@@ -66,6 +80,7 @@ public class MedicationRequestPlanMapperTest {
     private static final int TWO = 2;
     private static final int SIX = 6;
     private static final int TWENTY_EIGHT = 28;
+    public static final String STATUS_REASON = "statusReason";
 
     @Mock
     private MedicationMapper medicationMapper;
@@ -81,16 +96,79 @@ public class MedicationRequestPlanMapperTest {
 
     @Test
     public void When_MappingAuthoriseResourceWithAllOptionals_Expect_AllFieldsToBeMappedCorrectly() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract2.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        assertThat(medicationStatement.isPresent()).isTrue();
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                        <effectiveTime>
+                            <high value="20060427"/>
+                        </effectiveTime>
+                        <availabilityTime value="20100114"/>
+                        <repeatNumber value="6"/>
+                        <quantity value="28" unit="1" />
+                        <predecessor typeCode="SUCC">
+                            <priorMedicationRef classCode="SBADM" moodCode="INT">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </predecessor>
+                        <pertinentInformation typeCode="PERT">
+                            <pertinentSupplyAnnotation classCode="OBS" moodCode="EVN">
+                                <text>Pharmacy Text: Repeat Dispensing Pharmacy Note. 1</text>
+                            </pertinentSupplyAnnotation>
+                        </pertinentInformation>
+                        <pertinentInformation typeCode="PERT">
+                            <pertinentSupplyAnnotation classCode="OBS" moodCode="EVN">
+                                <text>Pharmacy Text: Repeat Dispensing Pharmacy Note. 2</text>
+                            </pertinentSupplyAnnotation>
+                        </pertinentInformation>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyPrescribe classCode="SPLY" moodCode="RQO">
+                        <id root="9B4B797A-D674-4362-B666-2ADC8551EEDA"/>
+                        <quantity value="1" unit="1" />
+                        <inFulfillmentOf typeCode="FLFS">
+                            <priorMedicationRef moodCode="INT">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </inFulfillmentOf>
+                    </ehrSupplyPrescribe>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <availabilityTime value="20060426"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                        <pertinentInformation typeCode="PERT">
+                            <pertinentSupplyAnnotation classCode="OBS" moodCode="EVN">
+                                <text>Patient no longer requires these</text>
+                            </pertinentSupplyAnnotation>
+                        </pertinentInformation>
+                    </ehrSupplyDiscontinue>
+                </component>
+                <pertinentInformation typeCode="PERT">
+                    <pertinentMedicationDosage classCode="SBADM" moodCode="RMD">
+                        <text>One To Be Taken Each Day</text>
+                    </pertinentMedicationDosage>
+                </pertinentInformation>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.get());
-        assertThat(supplyAuthorise.isPresent()).isTrue();
-
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.get(),
-                PRACTISE_CODE);
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+                ehrExtract,
+                medicationStatement,
+                supplyAuthorise,
+                PRACTISE_CODE
+        );
 
         var repeatInformation = medicationRequest.getExtensionsByUrl(REPEAT_INFO_URL);
         assertThat(repeatInformation).hasSize(ONE);
@@ -120,204 +198,487 @@ public class MedicationRequestPlanMapperTest {
 
     @Test
     public void When_MappingAuthoriseResourceWithNoEffectiveTime_Expect_NoExpiryDateExtensionAdded() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract5.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        assertThat(medicationStatement.isPresent()).isTrue();
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                    </ehrSupplyAuthorise>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.get());
-        assertThat(supplyAuthorise.isPresent()).isTrue();
-
-        var medicationRequest
-            = medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.get(),
-            PRACTISE_CODE);
-
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
         var repeatInformation = medicationRequest.getExtensionsByUrl(REPEAT_INFO_URL);
-        assertThat(repeatInformation).hasSize(ONE);
-
         var expiryDate = repeatInformation.get(0).getExtensionsByUrl(REPEATS_EXPIRY_DATE_URL);
-        assertThat(expiryDate.size()).isZero();
+
+        assertAll(
+            () -> assertThat(repeatInformation).hasSize(ONE),
+            () -> assertThat(expiryDate.size()).isZero()
+        );
     }
 
     @Test
     public void When_MappingAuthoriseResourceEffectiveTimeWithNullHighValue_Expect_NoExpiryDateExtensionAdded() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract6.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        assertThat(medicationStatement.isPresent()).isTrue();
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                        <effectiveTime>
+                            <high nullFlavor="UNK" />
+                        </effectiveTime>
+                    </ehrSupplyAuthorise>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.get());
-        assertThat(supplyAuthorise.isPresent()).isTrue();
-
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.get(),
-                PRACTISE_CODE);
-
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
         var repeatInformation = medicationRequest.getExtensionsByUrl(REPEAT_INFO_URL);
-        assertThat(repeatInformation).hasSize(ONE);
-
         var expiryDate = repeatInformation.get(0).getExtensionsByUrl(REPEATS_EXPIRY_DATE_URL);
-        assertThat(expiryDate.size()).isZero();
+
+        assertAll(
+            () -> assertThat(repeatInformation).hasSize(ONE),
+            () -> assertThat(expiryDate.size()).isZero()
+        );
     }
 
     @Test
-    public void When_MappingDiscontinue_With_PertinentInformation_Expect_StatusReasonAdded() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract2.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+    public void When_MappingDiscontinueWithPertinentInformation_Expect_StatusReasonAdded() {
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <availabilityTime value="20060426"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                        <pertinentInformation typeCode="PERT">
+                            <pertinentSupplyAnnotation classCode="OBS" moodCode="EVN">
+                                <text>Patient no longer requires these</text>
+                            </pertinentSupplyAnnotation>
+                        </pertinentInformation>
+                    </ehrSupplyDiscontinue>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
-
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
         var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
-        assertThat(statusExt).hasSize(1);
-
-        var statusReasonExt = statusExt.get(0).getExtensionsByUrl("statusReason");
-        assertThat(statusReasonExt).hasSize(1);
-
+        var statusReasonExt = statusExt.get(0).getExtensionsByUrl(STATUS_REASON);
         var statusReason = (CodeableConcept) statusReasonExt.get(0).getValue();
 
-        assertThat(statusReason.getText()).isEqualTo("Patient no longer requires these");
+        assertAll(
+            () -> assertThat(statusExt).hasSize(1),
+            () -> assertThat(statusReasonExt).hasSize(1),
+            () -> assertThat(statusReason.getText()).isEqualTo("Patient no longer requires these")
+        );
     }
 
     @Test
-    public void When_MappingDiscontinue_With_MissingPertinentInformationAndCodeDisplayPresent_Expect_DefaultTextAddedAsReason() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract7.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+    public void When_MappingDiscontinueWithCodeDisplayAndMissingPertinentInformation_Expect_DefaultTextAddedAsReason() {
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <code code="EMISDRUG_DISCONTINUATION"
+                              codeSystem="2.16.840.1.113883.2.1.6.3"
+                              displayName="Medication Course Ended"
+                        />
+                        <availabilityTime value="20060426"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                    </ehrSupplyDiscontinue>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
-
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
         var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
-        assertThat(statusExt).hasSize(1);
-
-        var statusReasonExt = statusExt.get(0).getExtensionsByUrl("statusReason");
-        assertThat(statusReasonExt).hasSize(1);
-
+        var statusReasonExt = statusExt.get(0).getExtensionsByUrl(STATUS_REASON);
         var statusReason = (CodeableConcept) statusReasonExt.get(0).getValue();
-        assertThat(statusReason.getText()).isEqualTo(DEFAULT_STATUS_REASON);
+
+        assertAll(
+            () -> assertThat(statusExt).hasSize(1),
+            () -> assertThat(statusReasonExt).hasSize(1),
+            () -> assertThat(statusReason.getText()).isEqualTo(DEFAULT_STATUS_REASON)
+        );
     }
 
     @Test
-    public void When_MappingDiscontinue_With_NoPertinentInformationAndHasCodeOriginalText_Expect_CorrectText() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract8.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+    public void When_MappingDiscontinue_With_NoPertinentInformationAndHasCodeOriginalText_Expect_OriginalTextAndDefaultText() {
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <code code="EMISDRUG_DISCONTINUATION" codeSystem="2.16.840.1.113883.2.1.6.3">
+                            <originalText>Ended</originalText>
+                        </code>
+                        <availabilityTime value="20060426"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                    </ehrSupplyDiscontinue>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
-
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
         var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
-        assertThat(statusExt).hasSize(1);
-
-        var statusReasonExt = statusExt.get(0).getExtensionsByUrl("statusReason");
-        assertThat(statusReasonExt).hasSize(1);
-
+        var statusReasonExt = statusExt.get(0).getExtensionsByUrl(STATUS_REASON);
         var statusReason = (CodeableConcept) statusReasonExt.get(0).getValue();
-        assertThat(statusReason.getText()).isEqualTo("(Ended) " + DEFAULT_STATUS_REASON);
+
+        assertAll(
+            () -> assertThat(statusExt).hasSize(1),
+            () -> assertThat(statusReasonExt).hasSize(1),
+            () -> assertThat(statusReason.getText()).isEqualTo("(Ended) " + DEFAULT_STATUS_REASON)
+        );
     }
 
     @Test
     public void When_MappingDiscontinue_With_MissingPertinentInformation_Expect_DefaultTextAddedAsReason() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract9.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <availabilityTime value="20060426"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                    </ehrSupplyDiscontinue>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
-
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
         var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
-        assertThat(statusExt).hasSize(1);
-
-        var statusReasonExt = statusExt.get(0).getExtensionsByUrl("statusReason");
-        assertThat(statusReasonExt).hasSize(1);
-
+        var statusReasonExt = statusExt.get(0).getExtensionsByUrl(STATUS_REASON);
         var statusReason = (CodeableConcept) statusReasonExt.get(0).getValue();
-        assertThat(statusReason.getText()).isEqualTo(DEFAULT_STATUS_REASON);
+
+        assertAll(
+            () -> assertThat(statusExt).hasSize(1),
+            () -> assertThat(statusReasonExt).hasSize(1),
+            () -> assertThat(statusReason.getText()).isEqualTo(DEFAULT_STATUS_REASON)
+        );
+    }
+
+    @Test void When_MappingDiscontinue_With_CodingOriginalTextAndDifferentPertinentInformation_Expect_BothDisplayed() {
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <code code="EMISDRUG_DISCONTINUATION" codeSystem="2.16.840.1.113883.2.1.6.3">
+                            <originalText>Ended</originalText>
+                        </code>
+                        <availabilityTime value="20060426"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                        <pertinentInformation typeCode="PERT">
+                            <pertinentSupplyAnnotation classCode="OBS" moodCode="EVN">
+                                <text>Patient no longer requires these</text>
+                            </pertinentSupplyAnnotation>
+                        </pertinentInformation>
+                    </ehrSupplyDiscontinue>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
+
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
+        var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
+        var statusReasonExt = statusExt.get(0).getExtensionsByUrl(STATUS_REASON);
+        var statusReason = (CodeableConcept) statusReasonExt.get(0).getValue();
+
+        assertAll(
+            () -> assertThat(statusExt).hasSize(1),
+            () -> assertThat(statusReasonExt).hasSize(1),
+            () -> assertThat(statusReason.getText()).isEqualTo("(Ended) Patient no longer requires these")
+        );
+    }
+
+    @Test void When_MappingDiscontinue_With_CodingOriginalTextAndSameTextPertinentInformation_Expect_DisplayedOnce() {
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <code code="EMISDRUG_DISCONTINUATION" codeSystem="2.16.840.1.113883.2.1.6.3">
+                            <originalText>Ended</originalText>
+                        </code>
+                        <availabilityTime value="20060426"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                        <pertinentInformation typeCode="PERT">
+                            <pertinentSupplyAnnotation classCode="OBS" moodCode="EVN">
+                                <text>Ended</text>
+                            </pertinentSupplyAnnotation>
+                        </pertinentInformation>
+                    </ehrSupplyDiscontinue>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
+
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
+        var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
+        var statusReasonExt = statusExt.get(0).getExtensionsByUrl(STATUS_REASON);
+        var statusReason = (CodeableConcept) statusReasonExt.get(0).getValue();
+
+        assertAll(
+            () -> assertThat(statusExt).hasSize(1),
+            () -> assertThat(statusReasonExt).hasSize(1),
+            () -> assertThat(statusReason.getText()).isEqualTo("Ended")
+        );
     }
 
     @Test
-    public void When_MappingAuthoriseResource_WithActiveStatusAndNoDiscontinue_Expect_ActiveStatusAndNoStatus() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract10.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+    public void When_MappingAuthoriseResource_WithActiveStatusAndNoDiscontinue_Expect_ActiveStatus() {
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                        <statusCode code="ACTIVE"/>
+                    </ehrSupplyAuthorise>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
 
         assertThat(medicationRequest.getStatus()).isEqualTo(ACTIVE);
     }
 
     @Test
     public void When_MappingAuthoriseResource_WithCompleteStatusAndNoDiscontinue_Expect_CompletedStatus() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract11.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                        <statusCode code="COMPLETE"/>
+                    </ehrSupplyAuthorise>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
 
         assertThat(medicationRequest.getStatus()).isEqualTo(COMPLETED);
     }
 
     @Test
     public void When_MappingAuthoriseResource_With_NoDiscontinue_Expect_NoStatusReasonExtension() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract11.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                        <statusCode code="COMPLETE"/>
+                    </ehrSupplyAuthorise>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
-
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
         var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
+
         assertThat(statusExt).isEmpty();
     }
 
     @Test
     public void When_MappingDiscontinue_With_UnknownDate_Expect_DiscontinueIgnored() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract12.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
+        var medicationStatementXml = """
+            <MedicationStatement xmlns="urn:hl7-org:v3" classCode="SBADM" moodCode="INT">
+                <id root="B4D70A6D-2EE4-41B6-B1FB-F9F0AD84C503"/>
+                <component typeCode="COMP">
+                    <ehrSupplyAuthorise classCode="SPLY" moodCode="INT">
+                        <id root="TEST_ID"/>
+                        <statusCode code="COMPLETE"/>
+                    </ehrSupplyAuthorise>
+                </component>
+                <component typeCode="COMP">
+                    <ehrSupplyDiscontinue classCode="SPLY" moodCode="RQO">
+                        <id root="D0BF39CA-E656-4322-879F-83EE6E688053"/>
+                        <availabilityTime nullFlavor="UNK"/>
+                        <reversalOf typeCode="REV">
+                            <priorMedicationRef classCode="SBADM" moodCode="ORD">
+                                <id root="TEST_ID"/>
+                            </priorMedicationRef>
+                        </reversalOf>
+                    </ehrSupplyDiscontinue>
+                </component>
+            </MedicationStatement>
+            """;
+        var ehrExtract = unmarshallEhrExtractFromMedicationRequestXml(medicationStatementXml);
+        var medicationStatement = extractMedicationStatement(ehrExtract);
+        var supplyAuthorise = extractSupplyAuthorise(medicationStatement);
 
-        var medicationRequest =
-            medicationRequestPlanMapper.mapToPlanMedicationRequest(ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(),
-                PRACTISE_CODE);
+        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
+            ehrExtract,
+            medicationStatement,
+            supplyAuthorise,
+            PRACTISE_CODE
+        );
+        var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
 
         assertThat(medicationRequest.getStatus()).isEqualTo(COMPLETED);
-        var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
         assertThat(statusExt).isEmpty();
     }
 
-    @Test void When_MappingDiscontinue_With_CodingOriginalTextAndPertinentInformation_Expect_CorrectText() {
-        var ehrExtract = unmarshallEhrExtract("ehrExtract13.xml");
-        Optional<RCMRMT030101UKMedicationStatement> medicationStatement = extractMedicationStatement(ehrExtract);
-        Optional<RCMRMT030101UKAuthorise> supplyAuthorise = extractSupplyAuthorise(medicationStatement.orElseThrow());
-
-        var medicationRequest = medicationRequestPlanMapper.mapToPlanMedicationRequest(
-                ehrExtract, medicationStatement.get(), supplyAuthorise.orElseThrow(), PRACTISE_CODE);
-
-        var statusExt = medicationRequest.getExtensionsByUrl(MEDICATION_STATUS_REASON_URL);
-        assertThat(statusExt).hasSize(1);
-
-        var statusReasonExt = statusExt.get(0).getExtensionsByUrl("statusReason");
-        assertThat(statusReasonExt).hasSize(1);
-
-        var statusReason = (CodeableConcept) statusReasonExt.get(0).getValue();
-        assertThat(statusReason.getText()).isEqualTo("(Ended) Patient no longer requires these");
-    }
-
-    private Optional<RCMRMT030101UKAuthorise> extractSupplyAuthorise(RCMRMT030101UKMedicationStatement medicationStatement) {
+    private RCMRMT030101UKAuthorise extractSupplyAuthorise(RCMRMT030101UKMedicationStatement medicationStatement) {
         return medicationStatement
             .getComponent()
             .stream()
             .map(RCMRMT030101UKComponent2::getEhrSupplyAuthorise)
-            .findFirst();
+            .findAny().orElseThrow();
     }
 
     private void assertStatusReasonInformation(Extension extension) {
@@ -325,7 +686,7 @@ public class MedicationRequestPlanMapperTest {
         var changeDateValue = (DateTimeType) changeDate.getValue();
         assertThat(changeDateValue.getValue()).isEqualTo(DateFormatUtil.parseToDateTimeType(AVAILABILITY_TIME).getValue());
 
-        var statusReason = extension.getExtensionsByUrl("statusReason").get(0);
+        var statusReason = extension.getExtensionsByUrl(STATUS_REASON).get(0);
         assertThat(statusReason.hasValue()).isTrue();
     }
 
@@ -344,7 +705,7 @@ public class MedicationRequestPlanMapperTest {
         assertThat(date).isEqualTo(DateFormatUtil.parseToDateTimeType("20060427").toString());
     }
 
-    private Optional<RCMRMT030101UKMedicationStatement> extractMedicationStatement(RCMRMT030101UKEhrExtract ehrExtract) {
+    private RCMRMT030101UKMedicationStatement extractMedicationStatement(RCMRMT030101UKEhrExtract ehrExtract) {
         return ehrExtract
             .getComponent()
             .stream()
@@ -355,11 +716,13 @@ public class MedicationRequestPlanMapperTest {
             .map(RCMRMT030101UKEhrComposition::getComponent)
             .flatMap(List::stream)
             .map(RCMRMT030101UKComponent4::getMedicationStatement)
-            .findFirst();
+            .findFirst()
+            .orElseThrow();
     }
 
     @SneakyThrows
-    private RCMRMT030101UKEhrExtract unmarshallEhrExtract(String fileName) {
-        return unmarshallFile(getFile("classpath:" + XML_RESOURCES_BASE + fileName), RCMRMT030101UK04EhrExtract.class);
+    private RCMRMT030101UKEhrExtract unmarshallEhrExtractFromMedicationRequestXml(String medicationRequestXml) {
+        var ehrExtractXml = EHR_EXTRACT_WRAPPER.replace("{{MedicationStatement}}", medicationRequestXml);
+        return unmarshallString(ehrExtractXml, RCMRMT030101UK04EhrExtract.class);
     }
 }